Move AppGetProductVersionString and AppGetChromiumVersionString in appshell_helpers
