+++ appshell/appshell_helpers.h
@@ -39,4 +39,10 @@ CefString AppGetDocumentsDirectory();
 // Returns the default CEF cache location
 CefString AppGetCachePath();
 
+// Returns a string containing the product and version (e.g. "Brackets/0.19.0.0")
+CefString AppGetProductVersionString();
+
+// Returns a string containing "Chrome/" appends with its version (e.g. "Chrome/29.0.1547.65")
+CefString AppGetChromiumVersionString();
+
 }  // namespace appshell
