+++ appshell/appshell_helpers_mac.mm
@@ -24,6 +24,7 @@
 #include "appshell/appshell_helpers.h"
 
 #include "include/cef_base.h"
+#include "include/cef_version.h"
 #include "config.h"
 #include <Cocoa/Cocoa.h>
 
@@ -100,6 +101,27 @@ CefString AppGetCachePath() {
     std::string cachePath = std::string(AppGetSupportDirectory()) + "/cef_data";
 
     return CefString(cachePath);
+}
+
+CefString AppGetProductVersionString() {
+    NSMutableString *s = [NSMutableString stringWithString:APP_NAME];
+    [s replaceOccurrencesOfString:@" "
+                       withString:@""
+                          options:NSLiteralSearch
+                            range:NSMakeRange(0, [s length])];
+    [s appendString:@"/"];
+    [s appendString:(NSString*)[[NSBundle mainBundle]
+                                objectForInfoDictionaryKey:(NSString *)kCFBundleVersionKey]];
+    CefString result = CefString([s UTF8String]);
+    return result;
+}
+
+CefString AppGetChromiumVersionString() {
+    NSMutableString *s = [NSMutableString stringWithFormat:@"Chrome/%d.%d.%d.%d",
+                             cef_version_info(2), cef_version_info(3),
+                             cef_version_info(4), cef_version_info(5)];
+    CefString result = CefString([s UTF8String]);
+    return result;
 }
 
 }  // namespace appshell
