+++ appshell/cefclient_gtk.cc
@@ -266,13 +266,3 @@ int main(int argc, char* argv[]) {
 
   return 0;
 }
-
-CefString AppGetProductVersionString() {
-  // TODO
-  return CefString("");
-}
-
-CefString AppGetChromiumVersionString() {
-  // TODO
-  return CefString("");
-}
